requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
requests==2.22.0
stripe==2.60.0
toml==0.9.6
urllib3==1.25.3
Werkzeug==0.15.5
python-dotenv==1.0.0
